설정/북마크/스터디 페이지 파일 분리
