+++ src/app/profile/bookmark/study/BookmarkStudy.scss
@@ -1,0 +1,101 @@
+@import '../../../../scss/mixins';
+
+@mixin paginationButton($color){
+    font-size: 16px;
+    line-height: 21px;
+    color:$color;
+}
+
+@include desktop{
+
+
+// studyList
+    .studyList{
+        display: flex;
+        flex-direction: row;
+        flex:1;
+
+        justify-content: space-between;
+
+        margin-top: 28px;
+        // background-color: red;
+    }
+    .studyItemContainer{
+        // width: 18.69%;
+        // height: 15.9%;
+        display: flex;
+        flex-direction: column;
+        background-color: blue;
+        border-radius: 10px;
+
+        flex:1;
+
+        // width: 220px;
+        // height: 220px;
+    }
+    .studyItemName{
+        font-size: 12px;
+        line-height: 17px;
+        color: #ffffff;
+        margin-top: 14px;
+        margin-left: 16px;
+    }
+    .studyItemTitle{
+        font-size: 24px;
+        color:#ffffff;
+        margin-top: 24px;
+        margin-left: 16px;
+        margin-right: 16px;
+        line-height: 35px;
+    }
+    .studyItemInfoContainer{
+        margin-left: 12px;
+        margin-right: 12px;
+        margin-top: 28px;
+
+        display: flex;
+        flex-direction: row;
+        justify-content: space-between;
+    }
+    .studyItemInfo{
+        font-size: 10px;
+        line-height: 15px;
+        color:#ffffff;
+
+    }
+    .tagContainer{
+        background-color: #ffffff;
+
+        display: flex;
+        flex-direction: row;
+        height: 48px;
+        align-items: center;
+
+        margin-top: 4px;
+        padding-left: 7px;
+
+        // box-shadow: none | x-position y-position blur spread color | inset | initial | inherit
+        box-shadow: 1px 2px 4px 0px rgba(95, 95, 95, 0.16);
+    }
+    .tagItemText{
+        color:#6f80ff;
+        font-size: 12px;
+        line-height: 10px;
+
+        padding: 4px 17px;
+        margin-right: 4px;
+
+        border: solid;
+        border-radius: 12px;
+        border-width: 1px;
+        border-color: #6f80ff;
+    }
+
+
+    button:active{
+        opacity: 0.3;
+    }
+}
+
+@include tablet{}
+@include mobile{}
